+++ src/expr/node/UpdateNode.js
@@ -72,7 +72,7 @@ UpdateNode.prototype.eval = function() {
             for (var i = 0, len = this.params.length; i < len; i++) {
                 var paramScope = paramScopes[i];
                 if (paramScope) {
-                    paramScope.set('end', size[i]);
+                    paramScope.set('end', size[i] - 1);
                 }
             }
         }